+++ Muvr/Entities/MRManagedExercisePlan.swift
@@ -1,0 +1,7 @@
+import Foundation
+import MuvrKit
+import CoreData
+
+class MRManagedExercisePlan : NSManagedObject {
+ 
+}
