--- Muvr/MRExerciseTableViewCell.xib
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="9531" systemVersion="15C50" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9529"/>
-        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="sMn-JS-eYs" customClass="MRExerciseTableViewCell" customModule="Muvr" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="448" height="88"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Exercise" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="KO6-xq-XeV">
-                    <rect key="frame" x="58" y="8" width="171" height="21"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="171" id="rLg-QI-T4r"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                    <color key="textColor" red="0.0" green="0.47843137250000001" blue="1" alpha="1" colorSpace="calibratedRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="8 repetitions" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Uah-en-1U5">
-                    <rect key="frame" x="58" y="28" width="171" height="16"/>
-                    <fontDescription key="fontDescription" type="system" weight="thin" pointSize="13"/>
-                    <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="easy-moderate" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="iPL-sE-A7B">
-                    <rect key="frame" x="58" y="44" width="171" height="16"/>
-                    <fontDescription key="fontDescription" type="system" weight="thin" pointSize="13"/>
-                    <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="15 kg" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="3ur-TD-jda">
-                    <rect key="frame" x="58" y="61" width="171" height="16"/>
-                    <fontDescription key="fontDescription" type="system" weight="thin" pointSize="13"/>
-                    <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-            </subviews>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
-            <constraints>
-                <constraint firstItem="KO6-xq-XeV" firstAttribute="leading" secondItem="Uah-en-1U5" secondAttribute="leading" id="C61-Hy-H3Y"/>
-                <constraint firstItem="iPL-sE-A7B" firstAttribute="top" secondItem="Uah-en-1U5" secondAttribute="bottom" id="CCd-5x-7W9"/>
-                <constraint firstItem="3ur-TD-jda" firstAttribute="trailing" secondItem="iPL-sE-A7B" secondAttribute="trailing" id="D0g-t6-UqN"/>
-                <constraint firstItem="Uah-en-1U5" firstAttribute="trailing" secondItem="iPL-sE-A7B" secondAttribute="trailing" id="DSi-r4-xTd"/>
-                <constraint firstItem="iPL-sE-A7B" firstAttribute="leading" secondItem="3ur-TD-jda" secondAttribute="leading" id="Rs6-ha-Usy"/>
-                <constraint firstItem="KO6-xq-XeV" firstAttribute="top" secondItem="sMn-JS-eYs" secondAttribute="topMargin" id="YnD-TY-Hdx"/>
-                <constraint firstItem="KO6-xq-XeV" firstAttribute="trailing" secondItem="Uah-en-1U5" secondAttribute="trailing" id="Zly-Ns-WWg"/>
-                <constraint firstItem="Uah-en-1U5" firstAttribute="top" secondItem="sMn-JS-eYs" secondAttribute="top" constant="28" id="hAp-Oc-2gA"/>
-                <constraint firstItem="iPL-sE-A7B" firstAttribute="leading" secondItem="Uah-en-1U5" secondAttribute="leading" id="i6R-pw-aPE"/>
-                <constraint firstItem="iPL-sE-A7B" firstAttribute="top" secondItem="KO6-xq-XeV" secondAttribute="bottom" constant="15" id="rBc-zY-T7L"/>
-                <constraint firstItem="3ur-TD-jda" firstAttribute="top" secondItem="iPL-sE-A7B" secondAttribute="bottom" constant="1" id="wN6-Ze-27Z"/>
-                <constraint firstItem="KO6-xq-XeV" firstAttribute="leading" secondItem="sMn-JS-eYs" secondAttribute="leading" constant="58" id="wQm-xS-PXr"/>
-            </constraints>
-            <nil key="simulatedStatusBarMetrics"/>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <connections>
-                <outlet property="exerciseLabel" destination="KO6-xq-XeV" id="KWy-al-cDu"/>
-                <outlet property="intensityLabel" destination="iPL-sE-A7B" id="Cqm-pw-gh9"/>
-                <outlet property="repetitionsLabel" destination="Uah-en-1U5" id="kh8-oE-99g"/>
-                <outlet property="weightLabel" destination="3ur-TD-jda" id="HTm-VA-7fH"/>
-            </connections>
-            <point key="canvasLocation" x="370" y="188"/>
-        </view>
-    </objects>
-</document>
