+++ Muvr/Session.storyboard
@@ -19,24 +19,12 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="736"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="UaB-VL-s9m" customClass="MRInteractivePlayerView" customModule="Muvr" customModuleProvider="target">
+                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="UaB-VL-s9m" customClass="MRExerciseView" customModule="Muvr" customModuleProvider="target">
                                 <rect key="frame" x="20" y="20" width="374" height="374"/>
-                                <color key="backgroundColor" white="0.66666666666666663" alpha="1" colorSpace="calibratedWhite"/>
+                                <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
                                 <constraints>
                                     <constraint firstAttribute="width" secondItem="UaB-VL-s9m" secondAttribute="height" multiplier="1:1" id="qCp-Hs-LGy"/>
                                 </constraints>
-                                <userDefinedRuntimeAttributes>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="coverImage" value="chart"/>
-                                    <userDefinedRuntimeAttribute type="number" keyPath="buttonSizes">
-                                        <real key="value" value="20"/>
-                                    </userDefinedRuntimeAttribute>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionOne_icon_selected" value="exercise"/>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionOne_icon_unselected" value="exercise_selected"/>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionTwo_icon_selected" value="exercise"/>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionTwo_icon_unselected" value="exercise_selected"/>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionThree_icon_selected" value="exercise"/>
-                                    <userDefinedRuntimeAttribute type="image" keyPath="actionThree_icon_unselected" value="exercise_selected"/>
-                                </userDefinedRuntimeAttributes>
                             </view>
                         </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
@@ -58,9 +46,4 @@
             <point key="canvasLocation" x="-2922" y="-456"/>
         </scene>
     </scenes>
-    <resources>
-        <image name="chart" width="649" height="671"/>
-        <image name="exercise" width="25" height="25"/>
-        <image name="exercise_selected" width="25" height="25"/>
-    </resources>
 </document>
